Nivora-Ver-loop-main/requirements.txt
# Optional dependencies for spotify_control.py
# Install with: pip install -r requirements.txt

# FREE TTS/STT Services (NEW)
edge-tts>=7.2.0     # Microsoft Edge TTS - FREE, no API key needed

# For volume control (Method 1)
pycaw>=20220416
comtypes>=1.1.11

# Browser Automation (NEW)
# Choose one or both:
playwright>=1.40.0  # Recommended - modern async API
selenium>=4.15.0    # Alternative - mature ecosystem

# Note: After installing playwright, run: python -m playwright install chromium

# Desktop Automation (Phase 2)
pyautogui>=0.9.54   # Screen automation - mouse/keyboard control
pywinauto>=0.6.8    # Windows UI Automation
pynput>=1.7.6       # Low-level keyboard/mouse control
pillow>=10.0.0      # Image processing

# File System (Phase 3)
aiofiles>=23.2.1    # Async file I/O
watchdog>=3.0.0     # File system monitoring

# Email & Calendar Enhancements (Phase 5)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
exchangelib>=5.0.0  # For Outlook/Exchange

# Notion Integration (Knowledge Management)
notion-client>=2.0.0

# Code Execution Sandbox (Phase 6)
docker>=6.1.3       # Docker SDK for Python
aiodocker>=0.21.0   # Async Docker client
psutil>=5.9.0       # Process management
